+++ Camara.ts
@@ -1,0 +1,6 @@
+
+class Camara {
+    public savarFoto(): string{
+        return "la foto del dia";
+    }
+}
